requirements.txt
line-bot-sdk
fastapi
uvicorn[standard]
langchain
faiss-cpu
tiktoken
python-dotenv
langchain-google-genai